[CORDA-1763]: Some review changes.
